+++ src/api/v1alpha1/renovatejob_types.go
@@ -15,17 +15,32 @@ import (
 // +kubebuilder:object:root=true
 // +kubebuilder:subresource:status
 type RenovateJobSpec struct {
-	Schedule        string                      `json:"schedule"`
-	Image           string                      `json:"image,omitempty"`
-	DiscoveryFilter string                      `json:"discoveryFilter"`
-	SecretRef       string                      `json:"secretRef,omitempty"`
-	ExtraEnv        []corev1.EnvVar             `json:"extraEnv,omitempty"`
-	Parallelism     int32                       `json:"parallelism"`
-	Resources       corev1.ResourceRequirements `json:"resources,omitempty"`
-	NodeSelector    map[string]string           `json:"nodeSelector,omitempty"`
-	ServiceAccount  *RenovateJobServiceAccount  `json:"serviceAccount,omitempty"`
-	Metadata        *RenovateJobMetadata        `json:"metadata,omitempty"`
+	// Cron schedule in standard cron format
+	Schedule string `json:"schedule"`
+	// Renovate Docker image to use
+	Image string `json:"image,omitempty"`
+	// Filter to select which projects to process
+	DiscoveryFilter string `json:"discoveryFilter,omitempty"`
+	// Topics to discover projects from
+	DiscoverTopics string `json:"discoverTopics,omitempty"`
+	// Reference to the secret containing the renovate config
+	SecretRef string `json:"secretRef,omitempty"`
+	// Additional environment variables to set in the renovate container
+	ExtraEnv []corev1.EnvVar `json:"extraEnv,omitempty"`
+	// Maximum number of projects to process in parallel
+	Parallelism int32 `json:"parallelism"`
+	// Resource requirements for the renovate container
+	Resources corev1.ResourceRequirements `json:"resources,omitempty"`
+	// Node selector for scheduling the resulting pod
+	NodeSelector map[string]string `json:"nodeSelector,omitempty"`
+	// Settings for the serviceaccount the renovate pod should use
+	ServiceAccount *RenovateJobServiceAccount `json:"serviceAccount,omitempty"`
+	// Metadata that shall be applied to the resulting pod
+	Metadata *RenovateJobMetadata `json:"metadata,omitempty"`
+	// Security context for the resulting pod and container
 	SecurityContext *RenovateJobSecurityContext `json:"securityContext,omitempty"`
+	// Configuration for webhooks to trigger renovate runs
+	Webhook *RenovateWebhook `json:"webhook,omitempty"`
 }
 
 // configuration regarding serviceaccounts for the resulting pod
@@ -40,12 +55,33 @@ type RenovateJobSecurityContext struct {
 	Container *corev1.SecurityContext    `json:"container,omitempty"`
 }
 
+// configuration for webhooks that can be used to trigger renovate runs
+type RenovateWebhook struct {
+	Enabled        bool                 `json:"enabled"`
+	Authentication *RenovateWebhookAuth `json:"authentication,omitempty"`
+}
+
+// authentication configuration for webhooks
+type RenovateWebhookAuth struct {
+	Enabled   bool                        `json:"enabled"`
+	SecretRef *RenovateSecretKeyReference `json:"secretRef,omitempty"`
+}
+
+// reference to a secret and key
+type RenovateSecretKeyReference struct {
+	Name string `json:"name,omitempty"`
+	Key  string `json:"key,omitempty"`
+}
+
 // metadata that shall be applied to the resulting pod
 type RenovateJobMetadata struct {
 	Labels      map[string]string `json:"labels,omitempty"`
 	Annotations map[string]string `json:"annotations,omitempty"`
 }
 
+/*
+Status of a single project within a RenovateJob
+*/
 type ProjectStatus struct {
 	Name    string                `json:"name"`
 	LastRun metav1.Time           `json:"lastRun"`
@@ -96,6 +132,7 @@ func (in *RenovateJob) ExecutorJobName(project string) string {
 	jobName := in.Name + "-" + project
 	jobName = strings.ReplaceAll(jobName, "/", "-") // Replace slashes to avoid issues with Kubernetes naming
 	jobName = strings.ReplaceAll(jobName, "_", "-")
+	jobName = strings.ReplaceAll(jobName, ".", "-")
 	jobName = strings.ToLower(jobName) // Ensure lowercase for consistency
 	return jobName
 }