+++ src/cmd/main.go
@@ -17,6 +17,7 @@ import (
 	"renovate-operator/internal/renovate"
 	"renovate-operator/scheduler"
 	"renovate-operator/ui"
+	"renovate-operator/webhook"
 
 	"k8s.io/client-go/rest"
 )
@@ -49,6 +50,23 @@ func main() {
 			},
 		},
 		{
+			Key:      "WEBHOOK_SERVER_PORT",
+			Optional: true,
+			Default:  "8082",
+			Validate: func(value string) error {
+				_, err := strconv.Atoi(value)
+				if err != nil {
+					return fmt.Errorf("'WEBHOOK_SERVER_PORT' needs to be an integer: %s", err.Error())
+				}
+				return nil
+			},
+		},
+		{
+			Key:      "WEBHOOK_SERVER_ENABLED",
+			Optional: true,
+			Default:  "false",
+		},
+		{
 			Key:      "DELETE_SUCCESSFULL_JOBS",
 			Optional: true,
 			Default:  "false",
@@ -61,6 +79,18 @@ func main() {
 				_, err := strconv.ParseInt(value, 10, 64)
 				if err != nil {
 					return fmt.Errorf("'JOB_TIMEOUT_SECONDS' needs to be an integer: %s", err.Error())
+				}
+				return nil
+			},
+		},
+		{
+			Key:      "JOB_BACKOFF_LIMIT",
+			Optional: true,
+			Default:  "1",
+			Validate: func(value string) error {
+				_, err := strconv.ParseInt(value, 10, 64)
+				if err != nil {
+					return fmt.Errorf("'JOB_BACKOFF_LIMIT' needs to be an integer: %s", err.Error())
 				}
 				return nil
 			},
@@ -110,8 +140,13 @@ func main() {
 	cronManager := scheduler.NewScheduler(ctrl.Log.WithName("scheduler"), health)
 	cronManager.Start()
 
-	uiServer := ui.NewServer(jobMgr, discovery, ctrl.Log.WithName("ui-server"), health)
+	uiServer := ui.NewServer(jobMgr, discovery, cronManager, ctrl.Log.WithName("ui-server"), health)
 	uiServer.Run()
+
+	if config.GetValue("WEBHOOK_SERVER_ENABLED") != "false" {
+		webhookServer := webhook.NewWebookServer(jobMgr, ctrl.Log.WithName("webhook"))
+		webhookServer.Run()
+	}
 
 	err = (&controllers.RenovateJobReconciler{
 		Scheduler: cronManager,