+++ src/internal/renovate/jobDefinitions.go
@@ -12,11 +12,23 @@ import (
 
 // create job spec for a discovery job
 func newDiscoveryJob(job *api.RenovateJob) *batchv1.Job {
-	predefinedEnvVars := []v1.EnvVar{}
+	predefinedEnvVars := []v1.EnvVar{
+		{
+			Name:  "NODE_NO_WARNINGS",
+			Value: "1",
+		},
+	}
+
 	if job.Spec.DiscoveryFilter != "" {
 		predefinedEnvVars = append(predefinedEnvVars, v1.EnvVar{
 			Name:  "RENOVATE_AUTODISCOVER_FILTER",
 			Value: job.Spec.DiscoveryFilter,
+		})
+	}
+	if job.Spec.DiscoverTopics != "" {
+		predefinedEnvVars = append(predefinedEnvVars, v1.EnvVar{
+			Name:  "RENOVATE_AUTODISCOVER_TOPICS",
+			Value: job.Spec.DiscoverTopics,
 		})
 	}
 
@@ -33,6 +45,7 @@ func newDiscoveryJob(job *api.RenovateJob) *batchv1.Job {
 	batchJob := &batchv1.Job{
 		Spec: batchv1.JobSpec{
 			ActiveDeadlineSeconds: getJobTimeoutSeconds(),
+			BackoffLimit:          getJobBackOffLimit(),
 			Template: v1.PodTemplateSpec{
 				Spec: v1.PodSpec{
 					ServiceAccountName: getServiceAccountName(job.Spec),
@@ -95,6 +108,7 @@ func newRenovateJob(job *api.RenovateJob, project string) *batchv1.Job {
 	batchJob := &batchv1.Job{
 		Spec: batchv1.JobSpec{
 			ActiveDeadlineSeconds: getJobTimeoutSeconds(),
+			BackoffLimit:          getJobBackOffLimit(),
 			Template: v1.PodTemplateSpec{
 				Spec: v1.PodSpec{
 					ServiceAccountName: getServiceAccountName(job.Spec),
@@ -196,4 +210,13 @@ func getJobTimeoutSeconds() *int64 {
 		return ptr.To(int64(1800))
 	}
 	return ptr.To(val)
+}
+
+func getJobBackOffLimit() *int32 {
+	timeoutString := config.GetValue("JOB_BACKOFF_LIMIT")
+	val, err := strconv.ParseInt(timeoutString, 10, 64)
+	if err != nil {
+		return ptr.To(int32(1800))
+	}
+	return ptr.To(int32(val))
 }